+++ src/Event/mt_evtMain.c
@@ -612,7 +612,7 @@ void FUN_00361070(int param_1);
 void FUN_00361140(int param_1,u16 *param_2);
 void FUN_00361200(int param_1,int param_2);
 void FUN_00361270(int param_1);
-void FUN_00361890(int *param_1,u64 param_2);
+void FUN_00361890(int *param_1,int param_2);
 void FUN_00361980(int param_1);
 void FUN_00361b60(int *param_1,int param_2);
 void FUN_00361c30(int param_1,int param_2);
@@ -1538,7 +1538,7 @@ int FUN_00361830(int param_1)
 // FUN_00361890 NONMATCHING
 
 
-void FUN_00361890(int *param_1,u64 param_2)
+void FUN_00361890(int *param_1,int param_2)
